factor out signal worker
